fix(review): title generation observability — changeset bump, function size, LogCapture dedupe
Address todo-287 review findings: the empty-frontmatter changeset was
discarded at version time despite shipping daemon behavior post
Rust-cutover, so it now bumps app-tauri patch. `do_generate_title` is
back under the 50-line limit by extracting `apply_generated_title`.
`LogCapture` moves to `mainframe-runtime` (already a shared dependency
of both crates) instead of being duplicated verbatim in mainframe-server.
