+++ packages/core-rs/crates/mainframe-chat/Cargo.toml
@@ -26,4 +26,3 @@ mainframe-background-tasks = { workspace = true }
 
 [dev-dependencies]
 tempfile = { workspace = true }
-tracing-subscriber = { workspace = true }
