+++ packages/core-rs/crates/mainframe-chat/src/test_support.rs
@@ -224,82 +224,9 @@ pub fn test_chat(id: &str) -> Chat {
     }
 }
 
-/// Captures `tracing` events emitted while a `#[tokio::test]` body runs, so
-/// title-generation logging tests can assert on structured fields instead of
-/// formatted message strings.
-///
-/// Install with `tracing::subscriber::set_default` (NOT `set_global_default`,
-/// which is process-wide and would fight parallel tests) and keep the guard
-/// alive for the duration of the awaited call. `#[tokio::test]` runs a
-/// current-thread runtime, so the thread-local dispatcher covers everything
-/// awaited in the test body — but a `tokio::spawn`ed task runs on its own
-/// dispatcher context and its events will NOT be captured.
-struct ReasonVisitor(Option<String>);
-
-impl tracing::field::Visit for ReasonVisitor {
-    fn record_str(&mut self, field: &tracing::field::Field, value: &str) {
-        if field.name() == "reason" {
-            self.0 = Some(value.to_string());
-        }
-    }
-    fn record_debug(&mut self, field: &tracing::field::Field, value: &dyn std::fmt::Debug) {
-        if field.name() == "reason" {
-            self.0 = Some(format!("{value:?}"));
-        }
-    }
-}
-
-/// One captured event: its level, and the `reason` field it carried (if any).
-type CapturedEvents = Arc<Mutex<Vec<(tracing::Level, Option<String>)>>>;
-
-pub struct LogCapture {
-    events: CapturedEvents,
-}
-
-impl<S: tracing::Subscriber> tracing_subscriber::Layer<S> for LogCapture {
-    fn on_event(
-        &self,
-        event: &tracing::Event<'_>,
-        _ctx: tracing_subscriber::layer::Context<'_, S>,
-    ) {
-        let mut visitor = ReasonVisitor(None);
-        event.record(&mut visitor);
-        self.events
-            .lock()
-            .unwrap_or_else(|e| e.into_inner())
-            .push((*event.metadata().level(), visitor.0));
-    }
-}
-
-impl LogCapture {
-    pub fn install() -> (impl tracing::Subscriber, CapturedEvents) {
-        use tracing_subscriber::layer::SubscriberExt;
-        let events = Arc::new(Mutex::new(Vec::new()));
-        let layer = LogCapture {
-            events: events.clone(),
-        };
-        (tracing_subscriber::registry().with(layer), events)
-    }
-
-    /// Captured events that carried a `reason` field, dropping any other
-    /// daemon logging so unrelated lines don't perturb the count.
-    pub fn events_with_reason(events: &CapturedEvents) -> Vec<(tracing::Level, String)> {
-        events
-            .lock()
-            .unwrap_or_else(|e| e.into_inner())
-            .iter()
-            .filter_map(|(level, reason)| reason.clone().map(|r| (*level, r)))
-            .collect()
-    }
-}
-
-#[test]
-fn log_capture_records_one_event_with_its_reason_field() {
-    let (subscriber, events) = LogCapture::install();
-    let _guard = tracing::subscriber::set_default(subscriber);
-    tracing::debug!(reason = "probe", "x");
-    let captured = LogCapture::events_with_reason(&events);
-    assert_eq!(captured, vec![(tracing::Level::DEBUG, "probe".to_string())]);
-}
+/// `mainframe-chat` and `mainframe-server` both depend on `mainframe-runtime`
+/// already, so the tracing capture helper lives there and is re-exported here
+/// for the crate's existing `crate::test_support::LogCapture` call sites.
+pub use mainframe_runtime::log_capture::LogCapture;
 
 // Not a port; test scaffolding only. No PORT STATUS trailer.
