+++ packages/core-rs/crates/mainframe-runtime/src/lib.rs
@@ -4,6 +4,7 @@
 
 pub mod auth;
 pub mod config;
+pub mod log_capture;
 pub mod logging;
 pub mod spawn_env;
 pub mod time;
