+++ packages/core-rs/crates/mainframe-server/Cargo.toml
@@ -46,4 +46,3 @@ reqwest = { workspace = true }
 hmac = { workspace = true }
 sha2 = { workspace = true }
 hex = { workspace = true }
-tracing-subscriber = { workspace = true }
